refactor: page -> segmentNode 수정
