+++ app/[locale]/community/news/create/page.tsx
@@ -11,10 +11,10 @@ import PageLayout from '@/components/layout/pageLayout/PageLayout';
 
 import { NEWS_TAGS } from '@/constants/tag';
 
-import { news } from '@/types/page';
 
 import { validateNewsForm } from '@/utils/formValidation';
 import { getPath } from '@/utils/page';
+import { news } from '@/utils/segmentNode';
 
 const newsPath = getPath(news);
