+++ app/[locale]/community/seminar/[id]/page.tsx
@@ -11,11 +11,11 @@ import { StraightNode } from '@/components/common/Nodes';
 import HTMLViewer from '@/components/editor/HTMLViewer';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 
-import { seminar } from '@/types/page';
 import { PostSearchQueryParams } from '@/types/post';
 
 import { getAdjPostsInfo } from '@/utils/getAdjPostInfo';
 import { getPath } from '@/utils/page';
+import { seminar } from '@/utils/segmentNode';
 
 interface SeminarPostPageProps {
   params: { id: string };
