+++ app/[locale]/reservations/introduction/page.tsx
@@ -3,10 +3,10 @@ import SelectionTitle from '@/components/common/selection/SelectionTitle';
 import HTMLViewer from '@/components/editor/HTMLViewer';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 
-import { reservationIntroduction } from '@/types/page';
 
 import { getPath } from '@/utils/page';
 import { replaceDashWithSpace } from '@/utils/replaceCharacter';
+import { reservationIntroduction } from '@/utils/segmentNode';
 
 const path = getPath(reservationIntroduction);
