+++ app/[locale]/reservations/privacy-policy/page.tsx
@@ -5,7 +5,7 @@ export const dynamic = 'force-dynamic';
 import HTMLViewer from '@/components/editor/HTMLViewer';
 import PageTitle from '@/components/layout/pageLayout/PageTitle';
 
-import { reservations } from '@/types/page';
+import { reservations } from '@/utils/segmentNode';
 
 export default function ReservationPrivacyPolicy() {
   return (
