+++ components/admin/important/ImportantListRow.tsx
@@ -2,9 +2,10 @@ import { Link } from '@/navigation';
 import CheckboxOrange from '@/public/image/checkbox_orange.svg';
 
 import { ImportantPostIdentifier, ImportantPreview } from '@/types/admin';
-import { news, notice, seminar } from '@/types/page';
 
 import { getPath } from '@/utils/page';
+import { news, notice, seminar } from '@/utils/segmentNode';
+
 
 interface ImportantListRowProps {
   index: number;
