+++ components/layout/navbar/NavbarRoot.tsx
@@ -6,10 +6,9 @@ import DotEmpty from '@/public/image/navbar/dot_empty.svg';
 import DotFill from '@/public/image/navbar/dot_fill.svg';
 import SnuLogo from '@/public/image/SNU_Logo.svg';
 
-import { SegmentNode, main as mainSegmentNode } from '@/types/page';
-
 import useCurrentSegmentNode from '@/utils/hooks/useCurrentSegmentNode';
 import { getPath, isAncestorNode } from '@/utils/page';
+import { SegmentNode, main as mainSegmentNode } from '@/utils/segmentNode';
 
 export default function NavbarRoot({
   state,
