+++ components/news/AdminFeatures.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link';
 
-import { news } from '@/types/page';
-
 import { getPath } from '@/utils/page';
+import { news } from '@/utils/segmentNode';
+
 
 const newsPath = getPath(news);
