+++ components/notice/AdminFeatures.tsx
@@ -2,10 +2,10 @@
 
 import { batchDeleteAction, batchUnpinAction } from '@/actions/noticeActions';
 
-import { notice } from '@/types/page';
 
 import useModal from '@/utils/hooks/useModal';
 import { getPath } from '@/utils/page';
+import { notice } from '@/utils/segmentNode';
 import { errorToast, successToast } from '@/utils/toast';
 
 import { BatchButton, CreateButton, EditButton } from './NoticeButtons';
