+++ components/research/labs/ResearchLabListRow.tsx
@@ -5,10 +5,10 @@ import { Fragment } from 'react';
 import { Link } from '@/navigation';
 import YoutubeIcon from '@/public/image/youtube_icon.svg';
 
-import { faculty, researchLabs } from '@/types/page';
 import { SimpleResearchLab } from '@/types/research';
 
 import { getPath } from '@/utils/page';
+import { faculty, researchLabs } from '@/utils/segmentNode';
 
 export const LAB_ROW_ITEM_WIDTH = {
   name: 'w-[14.5rem]',
