+++ components/research/labs/ResesarchLabInfo.tsx
@@ -6,10 +6,10 @@ import CornerFoldedRectangle from '@/components/common/CornerFoldedRectangle';
 
 import { COLOR_THEME } from '@/constants/color';
 
-import { faculty } from '@/types/page';
 import { ResearchLab } from '@/types/research';
 
 import { getPath } from '@/utils/page';
+import { faculty } from '@/utils/segmentNode';
 
 export default function ResearchLabInfo({ lab }: { lab: ResearchLab }) {
   const dropShadow = 'drop-shadow(1px 2px 2px rgba(0,0,0,0.25)';
