+++ components/search/NoticeRow.tsx
@@ -1,8 +1,8 @@
 import { Link } from '@/navigation';
 
-import { notice } from '@/types/page';
-
 import { getPath } from '@/utils/page';
+import { notice } from '@/utils/segmentNode';
+
 
 interface NoticeRowProps {
   id: number;
