+++ components/seminar/AdminFeatures.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link';
 
-import { seminar } from '@/types/page';
-
 import { getPath } from '@/utils/page';
+import { seminar } from '@/utils/segmentNode';
+
 
 const seminarPath = getPath(seminar);
