+++ utils/segmentNode.ts
